Auth by Owners email

+++ SourceCode/PantryBackend/Pantry.Recipe.Api/Endpoints/RecipeEndpoint.cs
@@ -4,6 +4,7 @@ using Microsoft.EntityFrameworkCore;
 using Pantry.Recipe.Api.Database.Contexts;
 using Pantry.Recipe.Api.Database.Entities;
 using Pantry.Services.RabbitMqServices;
+using Pantry.Services.UserServices;
 using Pantry.Shared.Models.MessageModes;
 using Pantry.Shared.Models.RecipeModels;
 using Pantry.Shared.Models.RecipeModels.RecipeRequestModels;
@@ -25,9 +26,12 @@ public static class RecipeEndpoint
         return group;
     }
 
-    private static async Task<IResult> UpdateRecipe(IMapper mapper, RecipeContext context, RecipeUpdateDto recipe)
+    private static async Task<IResult> UpdateRecipe(IMapper mapper, IHeaderEMailService eMailService, RecipeContext context, RecipeUpdateDto recipe)
     {
-        if (await context.Recipes.FindAsync(recipe.Id) is RecipeEntity recipeRecord)
+        var eMail = eMailService.GetHeaderEMail();
+        if (string.IsNullOrEmpty(eMail)) { return Results.Unauthorized(); }
+
+        if (await context.Recipes.FindAsync(recipe.Id) is RecipeEntity recipeRecord && recipeRecord.Owner == eMail)
         {
             recipeRecord.Name = recipe.Name;
             recipeRecord.Description = recipe.Description;
@@ -40,9 +44,12 @@ public static class RecipeEndpoint
         return Results.NotFound();
     }
 
-    private static async Task<IResult> DeleteRecipe(IRabbitMqPublisher rabbitMqPublisher, RecipeContext context, Guid id)
+    private static async Task<IResult> DeleteRecipe(IRabbitMqPublisher rabbitMqPublisher, IHeaderEMailService eMailService, RecipeContext context, Guid id)
     {
-        if (await context.Recipes.FindAsync(id) is RecipeEntity recipe)
+        var eMail = eMailService.GetHeaderEMail();
+        if (string.IsNullOrEmpty(eMail)) { return Results.Unauthorized(); }
+
+        if (await context.Recipes.FindAsync(id) is RecipeEntity recipe && recipe.Owner == eMail)
         {
             context.Recipes.Remove(recipe);
             await context.SaveChangesAsync();
@@ -55,19 +62,26 @@ public static class RecipeEndpoint
         return Results.NotFound();
     }
 
-    private static async Task<IResult> GetRecipe(IMapper mapper, RecipeContext context, Guid id)
+    private static async Task<IResult> GetRecipe(IMapper mapper, IHeaderEMailService eMailService, RecipeContext context, Guid id)
     {
-        return await context.Recipes.FindAsync(id) is RecipeEntity recipe
+        var eMail = eMailService.GetHeaderEMail();
+        if (string.IsNullOrEmpty(eMail)) { return Results.Unauthorized(); }
+
+        return await context.Recipes.FindAsync(id) is RecipeEntity recipe && recipe.Owner == eMail
             ? Results.Ok(mapper.Map<Shared.Models.RecipeModels.Recipe>(recipe))
             : Results.NotFound();
     }
 
-    private static async Task<IResult> CreateRecipe(IMapper mapper, RecipeContext context, RecipeCreateDto recipe)
+    private static async Task<IResult> CreateRecipe(IMapper mapper, IHeaderEMailService eMailService, RecipeContext context, RecipeCreateDto recipe)
     {
+        var eMail = eMailService.GetHeaderEMail();
+        if (string.IsNullOrEmpty(eMail)) { return Results.Unauthorized(); }
+
         var entity = mapper.Map<RecipeEntity>(recipe);
 
         entity.Id = Guid.NewGuid();
         entity.CreatedOn = DateTime.Now;
+        entity.Owner = eMail;
 
         await context.Recipes.AddAsync(entity);
         await context.SaveChangesAsync();
@@ -76,10 +90,13 @@ public static class RecipeEndpoint
     }
 
     
-    private static async Task<IResult> GetRecipeOverviewById(IMapper mapper, RecipeContext context, Guid id) {
+    private static async Task<IResult> GetRecipeOverviewById(IMapper mapper, IHeaderEMailService eMailService, RecipeContext context, Guid id) {
+        var eMail = eMailService.GetHeaderEMail();
+        if (string.IsNullOrEmpty(eMail)) { return Results.Unauthorized(); }
+
         var recipes = await context.Recipes.FindAsync(id);
 
-        if (await context.Recipes.FindAsync(id) is RecipeEntity recipe) {
+        if (await context.Recipes.FindAsync(id) is RecipeEntity recipe && recipe.Owner == eMail) {
             if (recipe.Description is { Length: > 50 })
                 recipe.Description = recipe.Description[..50];
             return Results.Ok(mapper.Map<RecipeOverview>(recipe));
@@ -90,8 +107,11 @@ public static class RecipeEndpoint
     }
 
 
-    private static async Task<IResult> GetRecipeOverview(IMapper mapper, RecipeContext context) {
-        var recipes = await context.Recipes.AsNoTracking().ToListAsync();
+    private static async Task<IResult> GetRecipeOverview(IMapper mapper, IHeaderEMailService eMailService, RecipeContext context) {
+        var eMail = eMailService.GetHeaderEMail();
+        if (string.IsNullOrEmpty(eMail)) { return Results.Unauthorized(); }
+
+        var recipes = await context.Recipes.AsNoTracking().Where(e => e.Owner == eMail).ToListAsync();
 
         foreach (var recipe in recipes) {
             if (recipe.Description is { Length: > 50 })
@@ -102,9 +122,12 @@ public static class RecipeEndpoint
     }
 
 
-    private static async Task<IResult> GetRecipes(IMapper mapper, RecipeContext context)
+    private static async Task<IResult> GetRecipes(IMapper mapper, IHeaderEMailService eMailService, RecipeContext context)
     {
-        var recipes = await context.Recipes.AsNoTracking().ToListAsync();
+        var eMail = eMailService.GetHeaderEMail();
+        if (string.IsNullOrEmpty(eMail)) { return Results.Unauthorized(); }
+
+        var recipes = await context.Recipes.AsNoTracking().Where(e => e.Owner == eMail).ToListAsync();
         return Results.Ok(mapper.Map<IEnumerable<Shared.Models.RecipeModels.Recipe>>(recipes));
     }
 }
