+++ SourceCode/PantryFrontend/pantry-app/src/routes/goods/[id]/+server.ts
@@ -1,14 +1,29 @@
 import { env } from '$env/dynamic/private';
+import { getSession } from '@auth/sveltekit';
+import { redirect } from '@sveltejs/kit';
 
-export async function DELETE({ params }): Promise<Response> {
+export async function DELETE({ params, request }): Promise<Response> {
+	const session = await getSession(request, { providers: [] });
+	if (!session?.user?.email) {
+		throw redirect(307, '/');
+	}
+
 	const response = await fetch(env.PRIVATE_PANTRY_API_URL + '/goods/' + params.id, {
-		method: 'DELETE'
+		method: 'DELETE',
+		headers: {
+			UserEMail: session?.user?.email
+		}
 	});
 
 	return response;
 }
 
 export async function POST({ request }): Promise<Response> {
+	const session = await getSession(request, { providers: [] });
+	if (!session?.user?.email) {
+		throw redirect(307, '/');
+	}
+
 	const { goodRatingCreateDto } = await request.json();
 	const response = await fetch(
 		env.PRIVATE_PANTRY_API_URL + '/goodratings/' + goodRatingCreateDto.GoodId,
@@ -16,7 +31,8 @@ export async function POST({ request }): Promise<Response> {
 			method: 'POST',
 			body: JSON.stringify(goodRatingCreateDto),
 			headers: {
-				'Content-Type': 'application/json'
+				'Content-Type': 'application/json',
+				UserEMail: session?.user?.email
 			}
 		}
 	);
