+++ SourceCode/PantryFrontend/pantry-app/src/routes/recipes/[id]/+server.ts
@@ -1,14 +1,29 @@
 import { env } from '$env/dynamic/private';
+import { getSession } from '@auth/sveltekit';
+import { redirect } from '@sveltejs/kit';
 
-export async function DELETE({ params }): Promise<Response> {
+export async function DELETE({ params, request }): Promise<Response> {
+	const session = await getSession(request, { providers: [] });
+	if (!session?.user?.email) {
+		throw redirect(307, '/');
+	}
+
 	const response = await fetch(env.PRIVATE_RECIPE_API_URL + '/recipes/' + params.id, {
-		method: 'DELETE'
+		method: 'DELETE',
+		headers: {
+			UserEMail: session?.user?.email
+		}
 	});
 
 	return response;
 }
 
 export async function POST({ request, params }): Promise<Response> {
+	const session = await getSession(request, { providers: [] });
+	if (!session?.user?.email) {
+		throw redirect(307, '/');
+	}
+
 	const { rating } = await request.json();
 
 	const response = await fetch(
@@ -17,7 +32,8 @@ export async function POST({ request, params }): Promise<Response> {
 			method: 'POST',
 			body: JSON.stringify(rating),
 			headers: {
-				'Content-Type': 'application/json'
+				'Content-Type': 'application/json',
+				UserEMail: session?.user?.email
 			}
 		}
 	);
@@ -25,7 +41,17 @@ export async function POST({ request, params }): Promise<Response> {
 	return response;
 }
 
-export async function GET({ params }): Promise<Response> {
-	const response = await fetch(env.PRIVATE_RECIPE_API_URL + '/recipes/overview/' + params.id);
+export async function GET({ params, request }): Promise<Response> {
+	const session = await getSession(request, { providers: [] });
+	if (!session?.user?.email) {
+		throw redirect(307, '/');
+	}
+
+	const response = await fetch(env.PRIVATE_RECIPE_API_URL + '/recipes/overview/' + params.id, {
+		method: 'GET',
+		headers: {
+			UserEMail: session?.user?.email
+		}
+	});
 	return response;
 }
